+++ app/factories/bulkrax/object_factory.rb
@@ -1,7 +1,6 @@
 # TODO: require 'importer/log_subscriber'
 module Bulkrax
   class ObjectFactory
-    include WithAssociatedCollection
     extend ActiveModel::Callbacks
     define_model_callbacks :save, :create
     class_attribute :system_identifier_field
@@ -156,6 +155,40 @@ module Bulkrax
       return attrs if attrs['collection_type_gid'].present?
       attrs['collection_type_gid'] = Hyrax::CollectionType.find_or_create_default_collection_type.gid 
       attrs
+    end
+
+    # Strip out the :collection key, and add the member_of_collection_ids,
+    # which is used by Hyrax::Actors::AddAsMemberOfCollectionsActor
+    def create_attributes
+      return transform_attributes if klass == Collection
+      if attributes[:collection].present?
+        transform_attributes.except(:collection).merge(member_of_collections_attributes: {0 => {id: collection.id}})
+      elsif attributes[:collections].present?
+        collection_ids = attributes[:collections].each.with_index.inject({}) do |ids, (element, index)|
+          ids[index] = element
+          ids
+        end
+        transform_attributes.except(:collections).merge(member_of_collections_attributes: collection_ids)
+      else
+        transform_attributes
+      end
+    end
+
+    # Strip out the :collection key, and add the member_of_collection_ids,
+    # which is used by Hyrax::Actors::AddAsMemberOfCollectionsActor
+    def update_attributes
+      return transform_attributes.except(:id) if klass == Collection
+      if attributes[:collection].present?
+        transform_attributes.except(:id).except(:collection).merge(member_of_collections_attributes: {0 => {id: collection.id}})
+      elsif attributes[:collections].present?
+        collection_ids = attributes[:collections].each.with_index.inject({}) do |ids, (element, index)|
+          ids[index] = element
+          ids
+        end
+        transform_attributes.except(:id).except(:collections).merge(member_of_collections_attributes: collection_ids)
+      else
+        transform_attributes.except(:id)
+      end
     end
 
     # Override if we need to map the attributes from the parser in
