+++ app/jobs/bulkrax/import_work_collection_job.rb
@@ -10,6 +10,7 @@ module Bulkrax
         ImporterRun.find(args[1]).increment!(:processed_collections)
       rescue => e
         ImporterRun.find(args[1]).increment!(:failed_collections)
+        raise e
       end
     end
   end
